Add two functions to check if text or file has a frontmatter
Although there is a `detect_format` function, which checks if there is
any handler that can check the file or string for a frontmatter, there
is no explicit function to check if a string or file has a frontmatter
without having to do the parse the file or string and check for metadata
set.

Thus, to make a frontmatter check on files or strings more explicit,
without necessarily parsing, two functions have been added: `check` and`
checks`, which use the `detect_format` function to do so.