Updates documentation of `check` and` checks` functions
In the creation of the tests for the created functions, it was seen that 
the documentation of the functions `check` and` checks` did not inform 
about the use case of having a frontmatter in the file or string but it 
is empty. Thus, the documentation of the functions has been changed to 
reflect the expected behavior.